move casemask helper functions to separate header
